requirements.txt
attrs==24.2.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
outcome==1.3.0.post0
PySocks==1.7.1
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
telebot==0.0.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
